Refactor: 서비스에서 반환 객체 엔티티로 변경 및 메소드명 변경
