+++ backend/src/main/java/com/api/backend/category/service/ScheduleCategoryService.java
@@ -1,20 +1,77 @@
 package com.api.backend.category.service;
 
-import com.api.backend.category.data.dto.ScheduleCategoryDto;
+import static com.api.backend.global.exception.type.ErrorCode.SCHEDULE_CATEGORY_NOT_FOUND_EXCEPTION;
+import static com.api.backend.global.exception.type.ErrorCode.TEAM_NOT_FOUND_EXCEPTION;
+
 import com.api.backend.category.data.dto.ScheduleCategoryEditRequest;
 import com.api.backend.category.data.dto.ScheduleCategoryRequest;
-import com.api.backend.category.data.dto.ScheduleCategoryResponse;
+import com.api.backend.category.data.entity.ScheduleCategory;
+import com.api.backend.category.data.repository.ScheduleCategoryRepository;
 import com.api.backend.category.type.CategoryType;
+import com.api.backend.global.exception.CustomException;
+import com.api.backend.team.data.entity.Team;
+import com.api.backend.team.data.repository.TeamRepository;
+import javax.transaction.Transactional;
+import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
 
-public interface ScheduleCategoryService {
 
-  ScheduleCategoryDto add(ScheduleCategoryRequest scheduleCategoryRequest, Long teamId);
+@Service
+@AllArgsConstructor
+public class ScheduleCategoryService {
 
-  Page<ScheduleCategoryResponse> searchByCategoryType(CategoryType categoryType, Pageable pageable);
+  private final ScheduleCategoryRepository scheduleCategoryRepository;
+  private final TeamRepository teamRepository;
 
-  ScheduleCategoryDto edit(ScheduleCategoryEditRequest scheduleCategoryEditRequest, Long teamId);
 
-  void delete(Long categoryId);
+  @Transactional
+  public ScheduleCategory add(ScheduleCategoryRequest scheduleCategoryRequest, Long teamId) {
+    Team team = validateTeam(teamId);
+
+    ScheduleCategory scheduleCategory = ScheduleCategory.builder()
+        .scheduleCategoryId(scheduleCategoryRequest.getCategoryId())
+        .team(team)
+        .categoryName(scheduleCategoryRequest.getCategoryName())
+        .categoryType(scheduleCategoryRequest.getCategoryType())
+        .color(scheduleCategoryRequest.getColor())
+        .build();
+    scheduleCategoryRepository.save(scheduleCategory);
+    return scheduleCategory;
+  }
+
+
+  public Page<ScheduleCategory> searchByCategoryType(CategoryType categoryType,
+      Pageable pageable) {
+    return scheduleCategoryRepository.findAllByCategoryType(categoryType, pageable);
+  }
+
+  @Transactional
+  public ScheduleCategory edit(ScheduleCategoryEditRequest scheduleCategoryEditRequest,
+      Long teamId) {
+    validateTeam(teamId);
+    ScheduleCategory scheduleCategory = validateScheduleCategory(
+        scheduleCategoryEditRequest.getCategoryId());
+    scheduleCategory.editScheduleCategory(scheduleCategoryEditRequest);
+    return scheduleCategoryRepository.save(scheduleCategory);
+  }
+
+  @Transactional
+  public void delete(Long categoryId) {
+    validateScheduleCategory(categoryId);
+    scheduleCategoryRepository.deleteById(categoryId);
+  }
+
+
+  public Team validateTeam(Long teamId) {
+    return teamRepository.findById(teamId)
+        .orElseThrow(() -> new CustomException(TEAM_NOT_FOUND_EXCEPTION));
+  }
+
+  public ScheduleCategory validateScheduleCategory(Long scheduleCategoryId) {
+    return scheduleCategoryRepository.findById(scheduleCategoryId)
+        .orElseThrow(() -> new CustomException(SCHEDULE_CATEGORY_NOT_FOUND_EXCEPTION));
+  }
+
 }
